requirement.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.8.0
fastapi==0.116.1
Flask==3.1.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
motor==3.7.1
packaging==25.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.15.4
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
